cache help info; cht locale

+++ check-i18n.py
@@ -3,6 +3,8 @@ import glob
 import os
 import re
 import shutil
+from opencc import OpenCC
+s2t = OpenCC('s2t').convert
 
 
 def load_json(json_src):
@@ -24,6 +26,17 @@ locales = {
 }
 
 
+CHS_KEY = 'src/locales/chs.json'
+CHT_KEY = 'src/locales/cht.json'
+
+if CHT_KEY not in locales:
+    locales[CHT_KEY] = {}
+
+for k in locales[CHS_KEY]:
+    if k not in locales[CHT_KEY]:
+        locales[CHT_KEY][k] = s2t(locales[CHS_KEY][k])
+
+
 for pwd, dirs, files in os.walk('src'):
     for file in files:
         if file.endswith('.vue'):
@@ -37,6 +50,7 @@ for pwd, dirs, files in os.walk('src'):
 
 
 for path, data in locales.items():
-    shutil.copy(path, path + '.bak')
+    if os.path.exists(path):
+        shutil.copy(path, path + '.bak')
     with open(path, 'w') as fout:
         json.dump(data, fout, ensure_ascii=False, indent=4)
